+++ src/gov/nih/nlm/mor/db/rxnorm/ConceptRelationship.java
@@ -1,28 +1,28 @@
 package gov.nih.nlm.mor.db.rxnorm;
 
 public class ConceptRelationship {
-	Integer id = null;
-	Integer conceptId1 = null;
+	Double id = null;
+	Double conceptId1 = null;
 	String relationship = "";
-	Integer conceptId2 = null;
+	Double conceptId2 = null;
 	
 	public ConceptRelationship() {
 		
 	}
 
-	public Integer getId() {
+	public Double getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Double id) {
 		this.id = id;
 	}
 
-	public Integer getConceptId1() {
+	public Double getConceptId1() {
 		return conceptId1;
 	}
 
-	public void setConceptId1(Integer conceptId1) {
+	public void setConceptId1(Double conceptId1) {
 		this.conceptId1 = conceptId1;
 	}
 
@@ -34,11 +34,11 @@ public class ConceptRelationship {
 		this.relationship = relationship;
 	}
 
-	public Integer getConceptId2() {
+	public Double getConceptId2() {
 		return conceptId2;
 	}
 
-	public void setConceptId2(Integer conceptId2) {
+	public void setConceptId2(Double conceptId2) {
 		this.conceptId2 = conceptId2;
 	}
 }
