+++ src/gov/nih/nlm/mor/db/rxnorm/TermRelationship.java
@@ -1,28 +1,28 @@
 package gov.nih.nlm.mor.db.rxnorm;
 
 public class TermRelationship {
-	Integer id = null;
-	Integer termId1 = null;
+	Double id = null;
+	Double termId1 = null;
 	String relationship = "";
-	Integer termId2 = null;
+	Double termId2 = null;
 	
 	public TermRelationship() {
 		
 	}
 
-	public Integer getId() {
+	public Double getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Double id) {
 		this.id = id;
 	}
 
-	public Integer getTermId1() {
+	public Double getTermId1() {
 		return termId1;
 	}
 
-	public void setTermId1(Integer termId1) {
+	public void setTermId1(Double termId1) {
 		this.termId1 = termId1;
 	}
 
@@ -34,11 +34,11 @@ public class TermRelationship {
 		this.relationship = relationship;
 	}
 
-	public Integer getTermId2() {
+	public Double getTermId2() {
 		return termId2;
 	}
 
-	public void setTermId2(Integer termId2) {
+	public void setTermId2(Double termId2) {
 		this.termId2 = termId2;
 	}
